refactor: move db and controller folder to src folder
