+++ src/app/route/usuarioRoute.js
@@ -1,5 +1,5 @@
 module.exports = function (app) {
-  var controller = new app.src.app.controller.usuario.UsuarioController(app);
+  var controller = new app.src.controller.usuario.UsuarioController(app);
   var util = new app.src.app.util.Util(app);
 
   app.get('/usuario', async function (req, res) {
