+++ src/controller/usuario/UsuarioController.js
@@ -1,9 +1,9 @@
-const UUIDGenerator = require('../../util/UUIDGenerator')
+const UUIDGenerator = require('../../app/util/UUIDGenerator')
 
 class UsuarioController {
   constructor(app) {
     this.app = app
-    this.repository = new this.app.src.app.db.repository.UsuarioRepository(this.app)
+    this.repository = new this.app.src.db.repository.UsuarioRepository(this.app)
   }
 
   async index (attributes, filter, order) {
